+++ src/hacks/index.ts
@@ -1,5 +1,7 @@
+import { enableDisableAppBackgroundBlur } from './disableAppBackgroundBlur.ts';
 import { enableEffectsCaching } from './effectsCaching.ts';
 import { enablePrecomputedNoiseTextures } from './precomputedNoiseTextures.ts';
+import { enableReducedLightingResolution } from './reduceLightingResolution.ts';
 import { enableSpritesheetSubstitution } from './spritesheetSubstitution.ts';
 import { enableTokenBarsCaching } from './tokenBarsCaching.ts';
 import { enableTokenRingSpritesheetSupport } from './tokenRingSpritesheetSupport.ts';
@@ -12,4 +14,6 @@ Hooks.once('setup', () => {
 	enableTokenRingSpritesheetSupport();
 	enableSpritesheetSubstitution();
 	enablePrecomputedNoiseTextures();
+	enableReducedLightingResolution();
+	enableDisableAppBackgroundBlur();
 });