+++ src/settings/constants.ts
@@ -4,6 +4,16 @@ export enum SETTINGS {
 	TokenBarsCaching = 'token-bars-caching',
 	SpritesheetSubstitution = 'spritesheet-substitution',
 	PrecomputedNoiseTextures = 'precomputed-noise-textures',
+	ReduceLightingResolution = 'reduce-lighting-resolution',
+	DisableAppV2BackgroundBlur = 'disable-app-v2-background-blur',
 	TokenRingSpritesheetSupport = 'token-ring-spritesheet-support',
 	CustomSpritesheets = 'spritesheet-substitution-custom',
+	CustomRenderScale = 'custom-render-scale',
 }
+
+export const RENDER_SCALE_DEFAULTS = {
+	background: 100,
+	illumination: 40,
+	coloration: 100,
+	darkness: 50,
+};